pyproject.toml
[project]
name = "atracker"
version = "0.1.0"
description = "Local-first activity watcher & tracker"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbus-next>=0.2.3",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "aiosqlite>=0.20.0",
    "pyyaml>=6.0.3",
]

[project.scripts]
atracker = "atracker.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/atracker"]
